# swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Alumni Influencers API",
        "description": "API documentation for the Alumni Influencers Bidding System",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/Alumni-Influencers",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/api/slots": {
            "get": {
                "tags": [
                    "Bidding System"
                ],
                "summary": "View available bidding slots",
                "operationId": "ccd8b255681ea257085c2b1a7ca3cde3",
                "responses": {
                    "200": {
                        "description": "List of slots"
                    },
                    "500": {
                        "description": "Error fetching slots"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/bids": {
            "put": {
                "tags": [
                    "Bidding System"
                ],
                "summary": "Update a bid amount",
                "operationId": "43305d75816e328d66f40d1615966217",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "bid_id",
                                    "bid_amount"
                                ],
                                "properties": {
                                    "bid_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "bid_amount": {
                                        "type": "number",
                                        "example": 60
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Bid updated"
                    },
                    "400": {
                        "description": "Invalid input"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Bidding System"
                ],
                "summary": "Place a new bid",
                "operationId": "f8f5bf0cb80e55972352a0da720f1c7e",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "slot_id",
                                    "bid_amount"
                                ],
                                "properties": {
                                    "slot_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "bid_amount": {
                                        "type": "number",
                                        "example": 50
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Bid placed"
                    },
                    "400": {
                        "description": "Invalid input or bid rejected"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Bidding System"
                ],
                "summary": "Cancel a bid",
                "operationId": "4ce90bf9b0069eddd88a7613071395e4",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "bid_id"
                                ],
                                "properties": {
                                    "bid_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Bid canceled"
                    },
                    "400": {
                        "description": "Invalid input"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/bidsstatus": {
            "get": {
                "tags": [
                    "Bidding System"
                ],
                "summary": "Get specific bid status",
                "operationId": "a93e21d10c89deaf6a95ed91bfc345df",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "bid_id"
                                ],
                                "properties": {
                                    "bid_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Status returned"
                    },
                    "400": {
                        "description": "Invalid bid_id"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/bids/history": {
            "get": {
                "tags": [
                    "Bidding System"
                ],
                "summary": "View alumni bidding history",
                "operationId": "8753d20f7462f948aa93a610627f1648",
                "responses": {
                    "200": {
                        "description": "History returned"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/cron/winner_selection": {
            "get": {
                "tags": [
                    "Cron Jobs"
                ],
                "summary": "Select winner for today's slot",
                "operationId": "86babc3b6f3eb39dd6877023cba1aac9",
                "parameters": [
                    {
                        "name": "cron_key",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Executed"
                    },
                    "403": {
                        "description": "Invalid cron key"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "apiKeyAuth": {
                "type": "apiKey",
                "name": "apikey",
                "in": "header"
            }
        }
    },
...
}